notatrix: Convert `src/utils/funcs.{j => t}s`
